requirements.txt
# Core App
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Vector Search
faiss-cpu>=1.7.4

# API / LLM
requests>=2.31.0

# File Handling
openpyxl>=3.1.2

# PDF Generation
reportlab>=4.0.0
